recovered r markdown file
After some troubleshooting, I was able to recover a previous auto-save of the r markdown file. I had to take a few minutes to rewrite some of the responses and fill in the work that was lost. I apologize that this is now being submitted after midnight, but if you are able to accept this for the 24hr late submission instead of the 48hr one, I would be very grateful.

A more in depth explanation was given in the blackboard submission of the assignment. Thank you for your understanding.